Modified the API proposal. Allow the model to passed in editor.openDialog call directly.
This will allow to set the model before calling dialog.show() method. Technically it could even be passed to the dialog constructor, however I don't think there will be a need to do that.

Temporarily I added model as an extra parameter in editor.openDialog - later on I'll turn the second argument of editor.openDialog into an object.

I can see now that there's a room for simplifying the change of this branch, which I'll do in subsequent commits.